+++ com-mmxlabs-common-util/com.mmxlabs.common.util/src/main/java/com/mmxlabs/common/curves/UnitCurve.java
@@ -12,7 +12,7 @@ public final class UnitCurve implements ICurve {
 	private static final UnitCurve INSTANCE = new UnitCurve();
 
 	@Override
-	public double getValueAtPoint(final double point) {
+	public int getValueAtPoint(final int point) {
 		return 1;
 	}
